requirements.txt
streamlit
farm-haystack[pdf]